Rename CyclicReferenceHandler as CyclicReferenceError for consistency
