+++ src/Page.js
@@ -7,7 +7,7 @@ const path = require('path');
 const pathIsInside = require('path-is-inside');
 const Promise = require('bluebird');
 
-const cyclicReferenceHandler = require('./lib/markbind/src/handlers/cyclicReferenceHandler.js');
+const CyclicReferenceError = require('./lib/markbind/src/handlers/cyclicReferenceError.js');
 const FsUtil = require('./util/fsUtil');
 const logger = require('./util/logger');
 const MarkBind = require('./lib/markbind/src/parser');
@@ -296,7 +296,7 @@ Page.prototype.collectHeadingsAndKeywordsInContent = function (content, lastHead
           try {
             this.collectHeadingsAndKeywordsInContent(includeContent, closestHeading, shouldExcludeHeadings);
           } catch (e) {
-            cyclicReferenceHandler.handle(e, includePath);
+            CyclicReferenceError.handle(e, includePath);
           }
         }
       } else {
